split popup html code
