+++ src/engine/render-index.js
@@ -1,13 +1,13 @@
-import {createProfileRatingTemplate} from "./view/index/profile-rating.js";
-import {createMainNavigationTemplate} from "./view/index/main-navigation.js";
-import {createSortTemplate} from "./view/index/sort.js";
-import {createFilmList} from "./view/index/film-list.js";
-import {createFilmCard} from "./view/index/film-card.js";
-import {createButtonShowMore} from "./view/index/button-show-more.js";
-import {createParagraph} from "./view/index/paragraph.js";
-import {generateFilm} from "./mocky/film.js";
-import {generateUser} from "./mocky/user.js";
-import {render} from "./utils.js";
+import {createProfileRatingTemplate} from "../view/index/profile-rating.js";
+import {createMainNavigationTemplate} from "../view/index/main-navigation.js";
+import {createSortTemplate} from "../view/index/sort.js";
+import {createFilmList} from "../view/index/film-list.js";
+import {createFilmCard} from "../view/index/film-card.js";
+import {createButtonShowMore} from "../view/index/button-show-more.js";
+import {createParagraph} from "../view/index/paragraph.js";
+import {generateFilm} from "../mocky/film.js";
+import {generateUser} from "../mocky/user.js";
+import {render} from "../utils";
 
 export const renderIndex = () => {
 
@@ -35,4 +35,4 @@ export const renderIndex = () => {
   render(siteFilmListElement, createButtonShowMore(), `beforeend`);
   render(siteFooterStaticsElement, createParagraph(), `beforeend`);
 
-  };
+};
